feed-generator/pyproject.toml
[tool.poetry]
name = "arg-monitoring"
version = "0.69.420"
description = "Monitors the ARG and stuff"
authors = ["clueless author <clueless@clueless.org>"]
readme = "../README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pytube = "^15.0.0"
requests = "^2.31.0"
dataclasses-json = "^0.6.4"
beautifulsoup4 = "^4.12.3"
feedgen = "^1.0.0"
pylint = "*"
mypy = "*"
flake8 = "*"
youtube-dl = { git = "https://github.com/ytdl-org/youtube-dl.git", branch = "master" }
pillow = "^10.2.0"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"
scikit-image = "^0.22.0"
matplotlib = "^3.8.3"
types-pillow = "^10.2.0.20240213"
twitchapi = "^4.2.0"
pygithub = "^2.2.0"
types-requests = "^2.31.0.20240218"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
